[JBIDE-14014] moved EmbeddableCartridge to com.openshift.client.cartridge
