+++ src/test/java/com/openshift/internal/client/EmbeddableCartridgeTest.java
@@ -28,11 +28,11 @@ import java.util.HashSet;
 import org.junit.Test;
 import org.mockito.Mockito;
 
-import com.openshift.client.EmbeddableCartridge;
 import com.openshift.client.IApplication;
 import com.openshift.client.IHttpClient;
 import com.openshift.client.IOpenShiftConnection;
 import com.openshift.client.OpenShiftConnectionFactory;
+import com.openshift.client.cartridge.EmbeddableCartridge;
 import com.openshift.client.cartridge.IEmbeddedCartridge;
 import com.openshift.client.utils.Samples;
 import com.openshift.internal.client.response.Link;
